% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_abc.R
\name{exp_abc}
\alias{exp_abc}
\title{Exponential ABC}
\usage{
exp_abc(
  true.lambda = 10,
  prior.lambda = c(1, 15),
  size = 10^3,
  nsim = 10^3,
  niter = 10^3,
  epsilon = 0.01
)
}
\arguments{
\item{true.lambda}{The true value of the parameter of lambda used to generate "observed" data (must be greater than 0)}

\item{prior.lambda}{The prior distribution of lambda used to generate the simulated data}

\item{size}{The size of the "observed" and simulated data generated}

\item{nsim}{The number of simulations to run/number of point estimates of lambda found}

\item{niter}{The number of iterations in ABC}

\item{epsilon}{Used to accept the posterior lambda values}
}
\value{
The point estimates for the posterior distribution parameter, lambda
}
\description{
Approximate Bayesian computation using Exponential distribution as the data generating
               mechanism. Finds point estimates for the posterior distribution parameter, lambda.
}
\examples{
exp_abc(true.lambda = 5, prior.lambda = c(1,10), size = 10^3, nsim = 100, niter = 10^3, epsilon = 0.01)
}
