% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gumbel_abc.R
\name{gumbel_abc}
\alias{gumbel_abc}
\title{Gumbel ABC}
\usage{
gumbel_abc(
  true.loc = 0,
  true.sc = 1.5,
  prior.loc = c(-5, 5),
  prior.sc = c(0.5, 3.5),
  size = 10^3,
  nsim = 10^3,
  niter = 10^4,
  epsilon = 0.01,
  plot = T
)
}
\arguments{
\item{true.loc}{The true value of the location parameter used to generate "observed" data}

\item{true.sc}{The true value of the scale parameter used to generate "observed" data (must be greater than 0)}

\item{prior.loc}{The prior distribution of the location parameter used to generate the simulated data}

\item{prior.sc}{The prior distribution of the scale parameter used to generate the simulated data}

\item{size}{The size of the "observed" and simulated data generated}

\item{nsim}{The number of simulations to run/number of point estimates of location and scale found}

\item{niter}{The number of iterations in ABC}

\item{epsilon}{Used to accept the posterior location and scale parameter values}
}
\value{
The point estimates for the posterior distribution parameters, location and scale
}
\description{
Approximate Bayesian computation using Gumbel distribution (light-tailed) as the data generating
               mechanism. Finds point estimates for the posterior distribution parameters, location and scale.
}
\examples{
gumbel_abc(true.loc = 1, true.sc = 2, prior.loc = c(-5,5), prior.sc = c(0.5,3.5), size = 10^3, nsim = 100, niter = 10^3, epsilon = 0.01)
}
